Move ContentModelEdit plugin to plugins package
