+++ packages-content-model/roosterjs-content-model-editor/lib/domUtils/eventUtils.ts
@@ -3,7 +3,6 @@ const ALT_CHAR_CODE = 'Alt';
 const META_CHAR_CODE = 'Meta';
 
 /**
- * @internal
  * Returns true when the event was fired from a modifier key, otherwise false
  * @param event The keyboard event object
  */
@@ -16,7 +15,6 @@ export function isModifierKey(event: KeyboardEvent): boolean {
 }
 
 /**
- * @internal
  * Returns true when the event was fired from a key that produces a character value, otherwise false
  * This detection is not 100% accurate. event.key is not fully supported by all browsers, and in some browsers (e.g. IE),
  * event.key is longer than 1 for num pad input. But here we just want to improve performance as much as possible.
