+++ packages-content-model/roosterjs-content-model-editor/lib/modelApi/edit/utils/deleteExpandedSelection.ts
@@ -1,9 +1,9 @@
 import { createInsertPoint } from '../utils/createInsertPoint';
-import { deleteBlock } from '../utils/deleteBlock';
-import { deleteSegment } from '../utils/deleteSegment';
+import { deleteBlock } from '../../../publicApi/block/deleteBlock';
+import { deleteSegment } from '../../../publicApi/segment/deleteSegment';
 import { iterateSelections } from '../../selection/iterateSelections';
+import type { DeleteSelectionContext } from '../../../publicTypes/parameter/DeleteSelectionStep';
 import type { ContentModelDocument } from 'roosterjs-content-model-types';
-import type { DeleteSelectionContext } from '../utils/DeleteSelectionStep';
 import type { FormatWithContentModelContext } from '../../../publicTypes/parameter/FormatWithContentModelContext';
 import type { IterateSelectionsOption } from '../../selection/iterateSelections';
 import {
