+++ packages-content-model/roosterjs-content-model-plugins/lib/index.ts
@@ -1,1 +1,2 @@
 export { ContentModelPastePlugin } from './paste/ContentModelPastePlugin';
+export { ContentModelEditPlugin } from './edit/ContentModelEditPlugin';
