+++ packages-content-model/roosterjs-content-model-plugins/test/edit/editingTestCommon.ts
@@ -1,9 +1,9 @@
 import { ContentModelDocument } from 'roosterjs-content-model-types';
-import { IContentModelEditor } from '../../../lib/publicTypes/IContentModelEditor';
 import {
+    IContentModelEditor,
     ContentModelFormatter,
     FormatWithContentModelOptions,
-} from '../../../lib/publicTypes/parameter/FormatWithContentModelContext';
+} from 'roosterjs-content-model-editor';
 
 export function editingTestCommon(
     apiName: string,
