+++ packages-content-model/roosterjs-content-model-plugins/test/edit/utils/getLeafSiblingBlockTest.ts
@@ -1,4 +1,4 @@
-import { getLeafSiblingBlock } from '../../../lib/modelApi/block/getLeafSiblingBlock';
+import { getLeafSiblingBlock } from '../../../lib/edit/utils/getLeafSiblingBlock';
 import {
     createContentModelDocument,
     createFormatContainer,
